Move to :core:data module.
