+++ Compose-Flight-Search-App/core/data/src/main/kotlin/com/example/flightsearchapp/core/data/datasource/AirportsDataSource.kt
@@ -1,13 +1,13 @@
-package com.example.flightsearchapp.data
+package com.example.flightsearchapp.core.data.datasource
 
-import com.example.flightsearchapp.data.database.AirportDao
-import com.example.flightsearchapp.data.database.AirportEntity
-import com.example.flightsearchapp.di.DispatcherIO
-import javax.inject.Inject
-import javax.inject.Singleton
+import com.example.flightsearchapp.core.data.di.DispatcherIO
+import com.example.flightsearchapp.core.database.dao.AirportDao
+import com.example.flightsearchapp.core.database.model.AirportEntity
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
+import javax.inject.Inject
+import javax.inject.Singleton
 
 interface AirportsDataSource {
     fun getAirportsByCodeOrNameStream(query: String): Flow<List<AirportEntity>>
