+++ Compose-Flight-Search-App/core/data/src/main/kotlin/com/example/flightsearchapp/core/data/di/CoroutineDispatchersModule.kt
@@ -1,4 +1,4 @@
-package com.example.flightsearchapp.di
+package com.example.flightsearchapp.core.data.di
 
 import dagger.Module
 import dagger.Provides
