+++ Compose-Flight-Search-App/core/data/src/main/kotlin/com/example/flightsearchapp/core/data/di/DataStoreModule.kt
@@ -1,4 +1,4 @@
-package com.example.flightsearchapp.di
+package com.example.flightsearchapp.core.data.di
 
 import android.content.Context
 import androidx.datastore.core.DataStore
@@ -13,10 +13,10 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
+import javax.inject.Singleton
 
 private const val SEARCH_TEXT = "search_text"
