+++ Compose-Flight-Search-App/core/data/src/main/kotlin/com/example/flightsearchapp/core/data/di/QualifierAnnotations.kt
@@ -1,4 +1,4 @@
-package com.example.flightsearchapp.di
+package com.example.flightsearchapp.core.data.di
 
 import javax.inject.Qualifier
