+++ Compose-Flight-Search-App/core/data/src/main/kotlin/com/example/flightsearchapp/core/data/model/FavoriteWithAirportsConverter.kt
@@ -1,7 +1,9 @@
-package com.example.flightsearchapp.ui.model
+package com.example.flightsearchapp.core.data.model
 
-fun com.example.flightsearchapp.database.model.FavoriteWithAirports.asFlight(): com.example.flightsearchapp.feature.searchscreen.ui.model.Flight =
-    com.example.flightsearchapp.feature.searchscreen.ui.model.Flight(
+import com.example.flightsearchapp.core.database.model.FavoriteWithAirports
+
+fun FavoriteWithAirports.asFlight(): Flight =
+    Flight(
         id = "${departureAirport.id}_${arriveAirport.id}",
         departureIataCode = departureAirport.iataCode,
         departureName = departureAirport.name,
